Merge repo

+++ lxd/storage_btrfs.go
@@ -2,11 +2,17 @@ package main
 
 import (
 	"fmt"
+	"io/ioutil"
 	"os"
 	"os/exec"
 	"path"
 	"path/filepath"
+	"strconv"
 	"strings"
+	"syscall"
+
+	"github.com/gorilla/websocket"
+	"github.com/pborman/uuid"
 
 	"github.com/lxc/lxd/shared"
 
@@ -45,7 +51,7 @@ func (s *storageBtrfs) Init(config map[string]interface{}) (storage, error) {
 }
 
 func (s *storageBtrfs) ContainerCreate(container container) error {
-	cPath := container.PathGet("")
+	cPath := container.Path()
 
 	// MkdirAll the pardir of the BTRFS Subvolume.
 	if err := os.MkdirAll(filepath.Dir(cPath), 0755); err != nil {
@@ -82,7 +88,7 @@ func (s *storageBtrfs) ContainerCreateFromImage(
 	}
 
 	// Now make a snapshot of the image subvol
-	err := s.subvolsSnapshot(imageSubvol, container.PathGet(""), false)
+	err := s.subvolsSnapshot(imageSubvol, container.Path(), false)
 	if err != nil {
 		return err
 	}
@@ -93,7 +99,7 @@ func (s *storageBtrfs) ContainerCreateFromImage(
 			return err
 		}
 	} else {
-		if err := os.Chmod(container.PathGet(""), 0700); err != nil {
+		if err := os.Chmod(container.Path(), 0700); err != nil {
 			return err
 		}
 	}
@@ -101,8 +107,12 @@ func (s *storageBtrfs) ContainerCreateFromImage(
 	return container.TemplateApply("create")
 }
 
+func (s *storageBtrfs) ContainerCanRestore(container container, sourceContainer container) error {
+	return nil
+}
+
 func (s *storageBtrfs) ContainerDelete(container container) error {
-	cPath := container.PathGet("")
+	cPath := container.Path()
 
 	// First remove the subvol (if it was one).
 	if s.isSubvolume(cPath) {
@@ -112,18 +122,20 @@ func (s *storageBtrfs) ContainerDelete(container container) error {
 	}
 
 	// Then the directory (if it still exists).
-	err := os.RemoveAll(cPath)
-	if err != nil {
-		s.log.Error("ContainerDelete: failed", log.Ctx{"cPath": cPath, "err": err})
-		return fmt.Errorf("Error cleaning up %s: %s", cPath, err)
+	if shared.PathExists(cPath) {
+		err := os.RemoveAll(cPath)
+		if err != nil {
+			s.log.Error("ContainerDelete: failed", log.Ctx{"cPath": cPath, "err": err})
+			return fmt.Errorf("Error cleaning up %s: %s", cPath, err)
+		}
 	}
 
 	return nil
 }
 
 func (s *storageBtrfs) ContainerCopy(container container, sourceContainer container) error {
-	subvol := sourceContainer.PathGet("")
-	dpath := container.PathGet("")
+	subvol := sourceContainer.Path()
+	dpath := container.Path()
 
 	if s.isSubvolume(subvol) {
 		// Snapshot the sourcecontainer
@@ -141,8 +153,8 @@ func (s *storageBtrfs) ContainerCopy(container container, sourceContainer contai
 		 * Copy by using rsync
 		 */
 		output, err := storageRsyncCopy(
-			sourceContainer.PathGet(""),
-			container.PathGet(""))
+			sourceContainer.Path(),
+			container.Path())
 		if err != nil {
 			s.ContainerDelete(container)
 
@@ -167,29 +179,34 @@ func (s *storageBtrfs) ContainerStop(container container) error {
 	return nil
 }
 
-func (s *storageBtrfs) ContainerRename(
-	container container, newName string) error {
-
-	oldPath := container.PathGet("")
-	newPath := container.PathGet(newName)
+func (s *storageBtrfs) ContainerRename(container container, newName string) error {
+	oldName := container.Name()
+	oldPath := container.Path()
+	newPath := containerPath(newName, false)
 
 	if err := os.Rename(oldPath, newPath); err != nil {
 		return err
 	}
 
-	// TODO: No TemplateApply here?
+	if shared.PathExists(shared.VarPath(fmt.Sprintf("snapshots/%s", oldName))) {
+		err := os.Rename(shared.VarPath(fmt.Sprintf("snapshots/%s", oldName)), shared.VarPath(fmt.Sprintf("snapshots/%s", newName)))
+		if err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
 
 func (s *storageBtrfs) ContainerRestore(
 	container container, sourceContainer container) error {
 
-	targetSubVol := container.PathGet("")
-	sourceSubVol := sourceContainer.PathGet("")
-	sourceBackupPath := container.PathGet("") + ".back"
+	targetSubVol := container.Path()
+	sourceSubVol := sourceContainer.Path()
+	sourceBackupPath := container.Path() + ".back"
 
 	// Create a backup of the container
-	err := os.Rename(container.PathGet(""), sourceBackupPath)
+	err := os.Rename(container.Path(), sourceBackupPath)
 	if err != nil {
 		return err
 	}
@@ -228,7 +245,7 @@ func (s *storageBtrfs) ContainerRestore(
 	if failure != nil {
 		// Restore original container
 		s.ContainerDelete(container)
-		os.Rename(sourceBackupPath, container.PathGet(""))
+		os.Rename(sourceBackupPath, container.Path())
 	} else {
 		// Remove the backup, we made
 		if s.isSubvolume(sourceBackupPath) {
@@ -240,11 +257,37 @@ func (s *storageBtrfs) ContainerRestore(
 	return failure
 }
 
+func (s *storageBtrfs) ContainerSetQuota(container container, size int64) error {
+	subvol := container.Path()
+
+	_, err := s.subvolQGroup(subvol)
+	if err != nil {
+		return err
+	}
+
+	output, err := exec.Command(
+		"btrfs",
+		"qgroup",
+		"limit",
+		"-e", fmt.Sprintf("%d", size),
+		subvol).CombinedOutput()
+
+	if err != nil {
+		return fmt.Errorf("Failed to set btrfs quota: %s", output)
+	}
+
+	return nil
+}
+
+func (s *storageBtrfs) ContainerGetUsage(container container) (int64, error) {
+	return s.subvolQGroupUsage(container.Path())
+}
+
 func (s *storageBtrfs) ContainerSnapshotCreate(
 	snapshotContainer container, sourceContainer container) error {
 
-	subvol := sourceContainer.PathGet("")
-	dpath := snapshotContainer.PathGet("")
+	subvol := sourceContainer.Path()
+	dpath := snapshotContainer.Path()
 
 	if s.isSubvolume(subvol) {
 		// Create a readonly snapshot of the source.
@@ -277,13 +320,49 @@ func (s *storageBtrfs) ContainerSnapshotDelete(
 
 	err := s.ContainerDelete(snapshotContainer)
 	if err != nil {
-		return fmt.Errorf("Error deleting snapshot %s: %s", snapshotContainer.NameGet(), err)
+		return fmt.Errorf("Error deleting snapshot %s: %s", snapshotContainer.Name(), err)
 	}
 
-	oldPathParent := filepath.Dir(snapshotContainer.PathGet(""))
+	oldPathParent := filepath.Dir(snapshotContainer.Path())
 	if ok, _ := shared.PathIsEmpty(oldPathParent); ok {
 		os.Remove(oldPathParent)
 	}
+	return nil
+}
+
+func (s *storageBtrfs) ContainerSnapshotStart(container container) error {
+	if shared.PathExists(container.Path() + ".ro") {
+		return fmt.Errorf("The snapshot is already mounted read-write.")
+	}
+
+	err := os.Rename(container.Path(), container.Path()+".ro")
+	if err != nil {
+		return err
+	}
+
+	err = s.subvolsSnapshot(container.Path()+".ro", container.Path(), false)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (s *storageBtrfs) ContainerSnapshotStop(container container) error {
+	if !shared.PathExists(container.Path() + ".ro") {
+		return fmt.Errorf("The snapshot isn't currently mounted read-write.")
+	}
+
+	err := s.subvolsDelete(container.Path())
+	if err != nil {
+		return err
+	}
+
+	err = os.Rename(container.Path()+".ro", container.Path())
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
 
@@ -291,8 +370,8 @@ func (s *storageBtrfs) ContainerSnapshotDelete(
 func (s *storageBtrfs) ContainerSnapshotRename(
 	snapshotContainer container, newName string) error {
 
-	oldPath := snapshotContainer.PathGet("")
-	newPath := snapshotContainer.PathGet(newName)
+	oldPath := snapshotContainer.Path()
+	newPath := containerPath(newName, true)
 
 	// Create the new parent.
 	if !shared.PathExists(filepath.Dir(newPath)) {
@@ -321,6 +400,11 @@ func (s *storageBtrfs) ContainerSnapshotRename(
 	return nil
 }
 
+func (s *storageBtrfs) ContainerSnapshotCreateEmpty(snapshotContainer container) error {
+	dpath := snapshotContainer.Path()
+	return s.subvolCreate(dpath)
+}
+
 func (s *storageBtrfs) ImageCreate(fingerprint string) error {
 	imagePath := shared.VarPath("images", fingerprint)
 	subvol := fmt.Sprintf("%s.btrfs", imagePath)
@@ -330,6 +414,7 @@ func (s *storageBtrfs) ImageCreate(fingerprint string) error {
 	}
 
 	if err := untarImage(imagePath, subvol); err != nil {
+		s.subvolDelete(subvol)
 		return err
 	}
 
@@ -340,7 +425,13 @@ func (s *storageBtrfs) ImageDelete(fingerprint string) error {
 	imagePath := shared.VarPath("images", fingerprint)
 	subvol := fmt.Sprintf("%s.btrfs", imagePath)
 
-	return s.subvolDelete(subvol)
+	if s.isSubvolume(subvol) {
+		if err := s.subvolsDelete(subvol); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 func (s *storageBtrfs) subvolCreate(subvol string) error {
@@ -371,7 +462,94 @@ func (s *storageBtrfs) subvolCreate(subvol string) error {
 	return nil
 }
 
+func (s *storageBtrfs) subvolQGroup(subvol string) (string, error) {
+	output, err := exec.Command(
+		"btrfs",
+		"qgroup",
+		"show",
+		subvol,
+		"-e",
+		"-f").CombinedOutput()
+
+	if err != nil {
+		return "", fmt.Errorf("btrfs quotas not supported. Try enabling them with 'btrfs quota enable'.")
+	}
+
+	var qgroup string
+	for _, line := range strings.Split(string(output), "\n") {
+		if line == "" || strings.HasPrefix(line, "qgroupid") || strings.HasPrefix(line, "---") {
+			continue
+		}
+
+		fields := strings.Fields(line)
+		if len(fields) != 4 {
+			continue
+		}
+
+		qgroup = fields[0]
+	}
+
+	if qgroup == "" {
+		return "", fmt.Errorf("Unable to find quota group")
+	}
+
+	return qgroup, nil
+}
+
+func (s *storageBtrfs) subvolQGroupUsage(subvol string) (int64, error) {
+	output, err := exec.Command(
+		"btrfs",
+		"qgroup",
+		"show",
+		subvol,
+		"-e",
+		"-f").CombinedOutput()
+
+	if err != nil {
+		return -1, fmt.Errorf("btrfs quotas not supported. Try enabling them with 'btrfs quota enable'.")
+	}
+
+	for _, line := range strings.Split(string(output), "\n") {
+		if line == "" || strings.HasPrefix(line, "qgroupid") || strings.HasPrefix(line, "---") {
+			continue
+		}
+
+		fields := strings.Fields(line)
+		if len(fields) != 4 {
+			continue
+		}
+
+		usage, err := strconv.ParseInt(fields[2], 10, 64)
+		if err != nil {
+			continue
+		}
+
+		return usage, nil
+	}
+
+	return -1, fmt.Errorf("Unable to find current qgroup usage")
+}
+
 func (s *storageBtrfs) subvolDelete(subvol string) error {
+	// Attempt (but don't fail on) to delete any qgroup on the subvolume
+	qgroup, err := s.subvolQGroup(subvol)
+	if err == nil {
+		output, err := exec.Command(
+			"btrfs",
+			"qgroup",
+			"destroy",
+			qgroup,
+			subvol).CombinedOutput()
+
+		if err != nil {
+			s.log.Warn(
+				"subvolume qgroup delete failed",
+				log.Ctx{"subvol": subvol, "output": string(output)},
+			)
+		}
+	}
+
+	// Delete the subvolume itself
 	output, err := exec.Command(
 		"btrfs",
 		"subvolume",
@@ -515,6 +693,32 @@ func (s *storageBtrfs) subvolsSnapshot(
  * else false.
  */
 func (s *storageBtrfs) isSubvolume(subvolPath string) bool {
+	if runningInUserns {
+		// subvolume show is restricted to real root, use a workaround
+
+		fs := syscall.Statfs_t{}
+		err := syscall.Statfs(subvolPath, &fs)
+		if err != nil {
+			return false
+		}
+
+		if fs.Type != filesystemSuperMagicBtrfs {
+			return false
+		}
+
+		parentFs := syscall.Statfs_t{}
+		err = syscall.Statfs(path.Dir(subvolPath), &parentFs)
+		if err != nil {
+			return false
+		}
+
+		if fs.Fsid == parentFs.Fsid {
+			return false
+		}
+
+		return true
+	}
+
 	output, err := exec.Command(
 		"btrfs",
 		"subvolume",
@@ -529,13 +733,43 @@ func (s *storageBtrfs) isSubvolume(subvolPath string) bool {
 
 // getSubVolumes returns a list of relative subvolume paths of "path".
 func (s *storageBtrfs) getSubVolumes(path string) ([]string, error) {
+	result := []string{}
+
+	if runningInUserns {
+		if !strings.HasSuffix(path, "/") {
+			path = path + "/"
+		}
+
+		// Unprivileged users can't get to fs internals
+		filepath.Walk(path, func(fpath string, fi os.FileInfo, err error) error {
+			if strings.TrimRight(fpath, "/") == strings.TrimRight(path, "/") {
+				return nil
+			}
+
+			if err != nil {
+				return nil
+			}
+
+			if !fi.IsDir() {
+				return nil
+			}
+
+			if s.isSubvolume(fpath) {
+				result = append(result, strings.TrimPrefix(fpath, path))
+			}
+			return nil
+		})
+
+		return result, nil
+	}
+
 	out, err := exec.Command(
 		"btrfs",
 		"inspect-internal",
 		"rootid",
 		path).CombinedOutput()
 	if err != nil {
-		return []string{}, fmt.Errorf(
+		return result, fmt.Errorf(
 			"Unable to get btrfs rootid, path='%s', err='%s'",
 			path,
 			err)
@@ -548,7 +782,7 @@ func (s *storageBtrfs) getSubVolumes(path string) ([]string, error) {
 		"subvolid-resolve",
 		rootid, path).CombinedOutput()
 	if err != nil {
-		return []string{}, fmt.Errorf(
+		return result, fmt.Errorf(
 			"Unable to resolve btrfs rootid, path='%s', err='%s'",
 			path,
 			err)
@@ -562,13 +796,12 @@ func (s *storageBtrfs) getSubVolumes(path string) ([]string, error) {
 		"-o",
 		path).CombinedOutput()
 	if err != nil {
-		return []string{}, fmt.Errorf(
+		return result, fmt.Errorf(
 			"Unable to list subvolumes, path='%s', err='%s'",
 			path,
 			err)
 	}
 
-	result := []string{}
 	lines := strings.Split(string(out), "\n")
 	for _, line := range lines {
 		if line == "" {
@@ -580,4 +813,265 @@ func (s *storageBtrfs) getSubVolumes(path string) ([]string, error) {
 	}
 
 	return result, nil
+}
+
+type btrfsMigrationSourceDriver struct {
+	container          container
+	snapshots          []container
+	btrfsSnapshotNames []string
+	btrfs              *storageBtrfs
+	runningSnapName    string
+	stoppedSnapName    string
+}
+
+func (s *btrfsMigrationSourceDriver) Snapshots() []container {
+	return s.snapshots
+}
+
+func (s *btrfsMigrationSourceDriver) send(conn *websocket.Conn, btrfsPath string, btrfsParent string) error {
+	args := []string{"send", btrfsPath}
+	if btrfsParent != "" {
+		args = append(args, "-p", btrfsParent)
+	}
+
+	cmd := exec.Command("btrfs", args...)
+
+	stdout, err := cmd.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	stderr, err := cmd.StderrPipe()
+	if err != nil {
+		return err
+	}
+
+	if err := cmd.Start(); err != nil {
+		return err
+	}
+
+	<-shared.WebsocketSendStream(conn, stdout)
+
+	output, err := ioutil.ReadAll(stderr)
+	if err != nil {
+		shared.Log.Error("problem reading btrfs send stderr", "err", err)
+	}
+
+	err = cmd.Wait()
+	if err != nil {
+		shared.Log.Error("problem with btrfs send", "output", string(output))
+	}
+	return err
+}
+
+func (s *btrfsMigrationSourceDriver) SendWhileRunning(conn *websocket.Conn) error {
+	if s.container.IsSnapshot() {
+		tmpPath := containerPath(fmt.Sprintf("%s/.migration-send-%s", s.container.Name(), uuid.NewRandom().String()), true)
+		err := os.MkdirAll(tmpPath, 0700)
+		if err != nil {
+			return err
+		}
+
+		btrfsPath := fmt.Sprintf("%s/.root", tmpPath)
+		if err := s.btrfs.subvolSnapshot(s.container.Path(), btrfsPath, true); err != nil {
+			return err
+		}
+
+		defer s.btrfs.subvolDelete(btrfsPath)
+
+		return s.send(conn, btrfsPath, "")
+	}
+
+	for i, snap := range s.snapshots {
+		prev := ""
+		if i > 0 {
+			prev = s.snapshots[i-1].Path()
+		}
+
+		if err := s.send(conn, snap.Path(), prev); err != nil {
+			return err
+		}
+	}
+
+	/* We can't send running fses, so let's snapshot the fs and send
+	 * the snapshot.
+	 */
+	tmpPath := containerPath(fmt.Sprintf("%s/.migration-send-%s", s.container.Name(), uuid.NewRandom().String()), true)
+	err := os.MkdirAll(tmpPath, 0700)
+	if err != nil {
+		return err
+	}
+
+	s.runningSnapName = fmt.Sprintf("%s/.root", tmpPath)
+	if err := s.btrfs.subvolSnapshot(s.container.Path(), s.runningSnapName, true); err != nil {
+		return err
+	}
+
+	btrfsParent := ""
+	if len(s.btrfsSnapshotNames) > 0 {
+		btrfsParent = s.btrfsSnapshotNames[len(s.btrfsSnapshotNames)-1]
+	}
+
+	return s.send(conn, s.runningSnapName, btrfsParent)
+}
+
+func (s *btrfsMigrationSourceDriver) SendAfterCheckpoint(conn *websocket.Conn) error {
+	tmpPath := containerPath(fmt.Sprintf("%s/.migration-send-%s", s.container.Name(), uuid.NewRandom().String()), true)
+	err := os.MkdirAll(tmpPath, 0700)
+	if err != nil {
+		return err
+	}
+
+	s.stoppedSnapName = fmt.Sprintf("%s/.root", tmpPath)
+	if err := s.btrfs.subvolSnapshot(s.container.Path(), s.stoppedSnapName, true); err != nil {
+		return err
+	}
+
+	return s.send(conn, s.stoppedSnapName, s.runningSnapName)
+}
+
+func (s *btrfsMigrationSourceDriver) Cleanup() {
+	if s.stoppedSnapName != "" {
+		s.btrfs.subvolDelete(s.stoppedSnapName)
+	}
+
+	if s.runningSnapName != "" {
+		s.btrfs.subvolDelete(s.runningSnapName)
+	}
+}
+
+func (s *storageBtrfs) MigrationType() MigrationFSType {
+	if runningInUserns {
+		return MigrationFSType_RSYNC
+	} else {
+		return MigrationFSType_BTRFS
+	}
+}
+
+func (s *storageBtrfs) MigrationSource(c container) (MigrationStorageSourceDriver, error) {
+	if runningInUserns {
+		return rsyncMigrationSource(c)
+	}
+
+	/* List all the snapshots in order of reverse creation. The idea here
+	 * is that we send the oldest to newest snapshot, hopefully saving on
+	 * xfer costs. Then, after all that, we send the container itself.
+	 */
+	snapshots, err := c.Snapshots()
+	if err != nil {
+		return nil, err
+	}
+
+	driver := &btrfsMigrationSourceDriver{
+		container:          c,
+		snapshots:          snapshots,
+		btrfsSnapshotNames: []string{},
+		btrfs:              s,
+	}
+
+	for _, snap := range snapshots {
+		btrfsPath := snap.Path()
+		driver.btrfsSnapshotNames = append(driver.btrfsSnapshotNames, btrfsPath)
+	}
+
+	return driver, nil
+}
+
+func (s *storageBtrfs) MigrationSink(live bool, container container, snapshots []container, conn *websocket.Conn) error {
+	if runningInUserns {
+		return rsyncMigrationSink(live, container, snapshots, conn)
+	}
+
+	cName := container.Name()
+
+	snapshotsPath := shared.VarPath(fmt.Sprintf("snapshots/%s", cName))
+	if !shared.PathExists(snapshotsPath) {
+		err := os.MkdirAll(shared.VarPath(fmt.Sprintf("snapshots/%s", cName)), 0700)
+		if err != nil {
+			return err
+		}
+	}
+
+	btrfsRecv := func(btrfsPath string, targetPath string, isSnapshot bool) error {
+		args := []string{"receive", "-e", btrfsPath}
+		cmd := exec.Command("btrfs", args...)
+
+		// Remove the existing pre-created subvolume
+		err := s.subvolsDelete(targetPath)
+		if err != nil {
+			return err
+		}
+
+		stdin, err := cmd.StdinPipe()
+		if err != nil {
+			return err
+		}
+
+		stderr, err := cmd.StderrPipe()
+		if err != nil {
+			return err
+		}
+
+		if err := cmd.Start(); err != nil {
+			return err
+		}
+
+		<-shared.WebsocketRecvStream(stdin, conn)
+
+		output, err := ioutil.ReadAll(stderr)
+		if err != nil {
+			shared.Debugf("problem reading btrfs receive stderr %s", err)
+		}
+
+		err = cmd.Wait()
+		if err != nil {
+			shared.Log.Error("problem with btrfs receive", log.Ctx{"output": string(output)})
+			return err
+		}
+
+		if !isSnapshot {
+			cPath := containerPath(fmt.Sprintf("%s/.root", cName), true)
+
+			err := s.subvolSnapshot(cPath, targetPath, false)
+			if err != nil {
+				shared.Log.Error("problem with btrfs snapshot", log.Ctx{"err": err})
+				return err
+			}
+
+			err = s.subvolsDelete(cPath)
+			if err != nil {
+				shared.Log.Error("problem with btrfs delete", log.Ctx{"err": err})
+				return err
+			}
+		}
+
+		return nil
+	}
+
+	for _, snap := range snapshots {
+		if err := btrfsRecv(containerPath(cName, true), snap.Path(), true); err != nil {
+			return err
+		}
+	}
+
+	/* finally, do the real container */
+	if err := btrfsRecv(containerPath(cName, true), container.Path(), false); err != nil {
+		return err
+	}
+
+	if live {
+		if err := btrfsRecv(containerPath(cName, true), container.Path(), false); err != nil {
+			return err
+		}
+	}
+
+	// Cleanup
+	if ok, _ := shared.PathIsEmpty(snapshotsPath); ok {
+		err := os.Remove(snapshotsPath)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
