+++ lxd/storage_dir.go
@@ -3,8 +3,11 @@ package main
 import (
 	"fmt"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
+
+	"github.com/gorilla/websocket"
 
 	"github.com/lxc/lxd/shared"
 
@@ -28,7 +31,7 @@ func (s *storageDir) Init(config map[string]interface{}) (storage, error) {
 }
 
 func (s *storageDir) ContainerCreate(container container) error {
-	cPath := container.PathGet("")
+	cPath := container.Path()
 	if err := os.MkdirAll(cPath, 0755); err != nil {
 		return fmt.Errorf("Error creating containers directory")
 	}
@@ -45,20 +48,20 @@ func (s *storageDir) ContainerCreate(container container) error {
 func (s *storageDir) ContainerCreateFromImage(
 	container container, imageFingerprint string) error {
 
-	rootfsPath := container.RootfsPathGet()
+	rootfsPath := container.RootfsPath()
 	if err := os.MkdirAll(rootfsPath, 0755); err != nil {
 		return fmt.Errorf("Error creating rootfs directory")
 	}
 
 	if container.IsPrivileged() {
-		if err := os.Chmod(container.PathGet(""), 0700); err != nil {
+		if err := os.Chmod(container.Path(), 0700); err != nil {
 			return err
 		}
 	}
 
 	imagePath := shared.VarPath("images", imageFingerprint)
-	if err := untarImage(imagePath, container.PathGet("")); err != nil {
-		os.RemoveAll(rootfsPath)
+	if err := untarImage(imagePath, container.Path()); err != nil {
+		s.ContainerDelete(container)
 		return err
 	}
 
@@ -72,13 +75,25 @@ func (s *storageDir) ContainerCreateFromImage(
 	return container.TemplateApply("create")
 }
 
+func (s *storageDir) ContainerCanRestore(container container, sourceContainer container) error {
+	return nil
+}
+
 func (s *storageDir) ContainerDelete(container container) error {
-	cPath := container.PathGet("")
+	cPath := container.Path()
+
+	if !shared.PathExists(cPath) {
+		return nil
+	}
 
 	err := os.RemoveAll(cPath)
 	if err != nil {
-		s.log.Error("ContainerDelete: failed", log.Ctx{"cPath": cPath, "err": err})
-		return fmt.Errorf("Error cleaning up %s: %s", cPath, err)
+		// RemovaAll fails on very long paths, so attempt an rm -Rf
+		output, err := exec.Command("rm", "-Rf", cPath).CombinedOutput()
+		if err != nil {
+			s.log.Error("ContainerDelete: failed", log.Ctx{"cPath": cPath, "output": output})
+			return fmt.Errorf("Error cleaning up %s: %s", cPath, string(output))
+		}
 	}
 
 	return nil
@@ -87,8 +102,8 @@ func (s *storageDir) ContainerDelete(container container) error {
 func (s *storageDir) ContainerCopy(
 	container container, sourceContainer container) error {
 
-	oldPath := sourceContainer.RootfsPathGet()
-	newPath := container.RootfsPathGet()
+	oldPath := sourceContainer.RootfsPath()
+	newPath := container.RootfsPath()
 
 	/*
 	 * Copy by using rsync
@@ -100,7 +115,7 @@ func (s *storageDir) ContainerCopy(
 		return fmt.Errorf("rsync failed: %s", string(output))
 	}
 
-	err = s.setUnprivUserAcl(sourceContainer, container.PathGet(""))
+	err = s.setUnprivUserAcl(sourceContainer, container.Path())
 	if err != nil {
 		return err
 	}
@@ -116,25 +131,31 @@ func (s *storageDir) ContainerStop(container container) error {
 	return nil
 }
 
-func (s *storageDir) ContainerRename(
-	container container, newName string) error {
+func (s *storageDir) ContainerRename(container container, newName string) error {
+	oldName := container.Name()
 
-	oldPath := container.PathGet("")
-	newPath := container.PathGet(newName)
+	oldPath := container.Path()
+	newPath := containerPath(newName, false)
 
 	if err := os.Rename(oldPath, newPath); err != nil {
 		return err
 	}
 
-	// TODO: No TemplateApply here?
+	if shared.PathExists(shared.VarPath(fmt.Sprintf("snapshots/%s", oldName))) {
+		err := os.Rename(shared.VarPath(fmt.Sprintf("snapshots/%s", oldName)), shared.VarPath(fmt.Sprintf("snapshots/%s", newName)))
+		if err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
 
 func (s *storageDir) ContainerRestore(
 	container container, sourceContainer container) error {
 
-	targetPath := container.PathGet("")
-	sourcePath := sourceContainer.PathGet("")
+	targetPath := container.Path()
+	sourcePath := sourceContainer.Path()
 
 	// Restore using rsync
 	output, err := storageRsyncCopy(
@@ -157,11 +178,19 @@ func (s *storageDir) ContainerRestore(
 	return nil
 }
 
+func (s *storageDir) ContainerSetQuota(container container, size int64) error {
+	return fmt.Errorf("The directory container backend doesn't support quotas.")
+}
+
+func (s *storageDir) ContainerGetUsage(container container) (int64, error) {
+	return -1, fmt.Errorf("The directory container backend doesn't support quotas.")
+}
+
 func (s *storageDir) ContainerSnapshotCreate(
 	snapshotContainer container, sourceContainer container) error {
 
-	oldPath := sourceContainer.PathGet("")
-	newPath := snapshotContainer.PathGet("")
+	oldPath := sourceContainer.Path()
+	newPath := snapshotContainer.Path()
 
 	/*
 	 * Copy by using rsync
@@ -177,14 +206,19 @@ func (s *storageDir) ContainerSnapshotCreate(
 
 	return nil
 }
+
+func (s *storageDir) ContainerSnapshotCreateEmpty(snapshotContainer container) error {
+	return os.MkdirAll(snapshotContainer.Path(), 0700)
+}
+
 func (s *storageDir) ContainerSnapshotDelete(
 	snapshotContainer container) error {
 	err := s.ContainerDelete(snapshotContainer)
 	if err != nil {
-		return fmt.Errorf("Error deleting snapshot %s: %s", snapshotContainer.NameGet(), err)
+		return fmt.Errorf("Error deleting snapshot %s: %s", snapshotContainer.Name(), err)
 	}
 
-	oldPathParent := filepath.Dir(snapshotContainer.PathGet(""))
+	oldPathParent := filepath.Dir(snapshotContainer.Path())
 	if ok, _ := shared.PathIsEmpty(oldPathParent); ok {
 		os.Remove(oldPathParent)
 	}
@@ -195,11 +229,11 @@ func (s *storageDir) ContainerSnapshotDelete(
 func (s *storageDir) ContainerSnapshotRename(
 	snapshotContainer container, newName string) error {
 
-	oldPath := snapshotContainer.PathGet("")
-	newPath := snapshotContainer.PathGet(newName)
+	oldPath := snapshotContainer.Path()
+	newPath := containerPath(newName, true)
 
 	// Create the new parent.
-	if strings.Contains(snapshotContainer.NameGet(), "/") {
+	if strings.Contains(snapshotContainer.Name(), "/") {
 		if !shared.PathExists(filepath.Dir(newPath)) {
 			os.MkdirAll(filepath.Dir(newPath), 0700)
 		}
@@ -211,12 +245,20 @@ func (s *storageDir) ContainerSnapshotRename(
 	}
 
 	// Remove the old parent (on container rename) if its empty.
-	if strings.Contains(snapshotContainer.NameGet(), "/") {
+	if strings.Contains(snapshotContainer.Name(), "/") {
 		if ok, _ := shared.PathIsEmpty(filepath.Dir(oldPath)); ok {
 			os.Remove(filepath.Dir(oldPath))
 		}
 	}
 
+	return nil
+}
+
+func (s *storageDir) ContainerSnapshotStart(container container) error {
+	return nil
+}
+
+func (s *storageDir) ContainerSnapshotStop(container container) error {
 	return nil
 }
 
@@ -226,4 +268,16 @@ func (s *storageDir) ImageCreate(fingerprint string) error {
 
 func (s *storageDir) ImageDelete(fingerprint string) error {
 	return nil
+}
+
+func (s *storageDir) MigrationType() MigrationFSType {
+	return MigrationFSType_RSYNC
+}
+
+func (s *storageDir) MigrationSource(container container) (MigrationStorageSourceDriver, error) {
+	return rsyncMigrationSource(container)
+}
+
+func (s *storageDir) MigrationSink(live bool, container container, snapshots []container, conn *websocket.Conn) error {
+	return rsyncMigrationSink(live, container, snapshots, conn)
 }
